// frontend/src/services/model/pickResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Football Pool API
 * OpenAPI spec version: 1.0.0
 */
import type { UserResponse } from "./userResponse";
import type { GameResponse } from "./gameResponse";

export interface PickResponse {
  id: number;
  user_id: number;
  user?: UserResponse;
  game_id: number;
  game?: GameResponse;
  picked: string;
  rank: number;
  quick_pick: boolean;
  created_at: string;
  updated_at: string;
}